requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain_google_genai==2.1.8
passlib==1.7.4
pydantic==2.11.7
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.42
